Change data type for storage of money and bonus credits to int

+++ OffCardAPP/src/main/java/clientAPI/BonusCreditStore.java
@@ -7,16 +7,16 @@ public interface BonusCreditStore {
 	/**
 	 *
 	 */
-	public void addBonusCredits(short amount) throws CardException;
+	public void addBonusCredits(int amount) throws CardException;
 
 	/**
 	 *
 	 */
-	public void removeBonusCredits(short amount) throws CardException;
+	public void removeBonusCredits(int amount) throws CardException;
 
 	/**
 	 *
 	 *
 	 */
-	public short checkBalance() throws CardException;
+	public int checkBalance() throws CardException;
 }
